requirements.txt
copilotagent>=0.1.14
cuteagent>=0.2.24
langchain>=1.0.0
langchain-anthropic>=1.0.0
langchain-core>=1.0.0
python-dotenv>=1.0.0
python-docx>=0.8.11
pandas>=2.0.0

